fix: surface real MinIO stat faults and map bad path/sort params to 400

+++ src/main/java/com/clara/ops/challenge/document_management_service_challenge/web/DocumentController.java
@@ -75,7 +75,7 @@ public class DocumentController {
             page.getSize(),
             page.getNumberOfElements(),
             page.getTotalPages(),
-            page.getTotalElements());
+            (int) page.getTotalElements());
     return new PaginatedDocumentSearch(meta, docs);
   }
